Move DOT/DAG support to common.
