+++ src/commonMain/kotlin/baaahs/ui/diagnostics/Dag.kt
@@ -7,42 +7,35 @@ import baaahs.gl.patch.ProgramNode
 import baaahs.gl.shader.InputPort
 import baaahs.show.live.LinkedPatch
 import baaahs.show.live.OpenPatch
-import external.dagre_d3.Graph
-import external.dagre_d3.d3
-import kotlinx.js.jso
 
-class Dag(
-    private val includePatchMods: Boolean = false
-) : ProgramVisitor() {
-    private var nextNode = 0
-    private val nodes = mutableMapOf<Any, String>()
+class DotDag(
+    includePatchMods: Boolean = false
+) : Dag(includePatchMods) {
     private val buf = StringBuilder()
-
-    val graph = Graph(jso { directed = true })
-        .also { it.setGraph(jso {}) }
     val text get() = buf.toString()
 
-    private fun declareNode(id: String, label: String, shape: String, style: String? = null) {
-        graph.setNode(id, jso {
-            this.label = label
-            this.shape = shape
-            this.style = style
-        })
+    override fun declareNode(id: String, label: String, shape: String, style: String?) {
         buf.append("    $id [${attrs(label, shape, style)}];\n")
+    }
+
+    override fun declareLink(fromId: String, toId: String, label: String) {
+        buf.append("    $fromId -> $toId [style=\"fill: none; stroke: #66f; stroke-width: 3px;\"];\n")
     }
 
     private fun attrs(label: String, shape: String, style: String?) =
         mapOf("label" to label, "shape" to shape, "style" to style)
-            .filterValues { it != null }.entries.map { (k, v) -> "$k=\"$v\"" }.joinToString(" ")
+            .filterValues { it != null }.entries.joinToString(" ") { (k, v) -> "$k=\"$v\"" }
+}
 
-    private fun declareLink(fromId: String, toId: String, label: String) {
-        graph.setEdge(fromId, toId, jso {
-            this.label = label
-            this.style = "fill: none; stroke: #66f; stroke-width: 3px; stroke-dasharray: 5, 5;"
-            this.curve = d3.curveBasis
-        })
-        buf.append("    $fromId -> $toId [style=\"fill: none; stroke: #66f; stroke-width: 3px;\"];\n")
-    }
+abstract class Dag(
+    private val includePatchMods: Boolean = false
+) : ProgramVisitor() {
+    private var nextNode = 0
+    private val nodes = mutableMapOf<Any, String>()
+
+    abstract fun declareNode(id: String, label: String, shape: String, style: String? = null)
+
+    abstract fun declareLink(fromId: String, toId: String, label: String)
 
     override fun visitFixtureType(fixtureType: FixtureType) {
         nodes.getOrPut(fixtureType) {
